refactor(bayn): modularize intent store
